+++ VoiceOver Designer/Features/Tests/EditorTests/Doubles/DocumentFake.swift
@@ -1,0 +1,12 @@
+import XCTest
+@testable import Editor
+import Combine
+import Document
+
+class DocumentFake: VODesignDocumentProtocol {
+    var controlsPublisher: PassthroughSubject<[A11yDescription], Never> = .init()
+    
+    var controls: [A11yDescription] = []
+    var undoManager: UndoManager? = nil
+    var image: Image? = nil
+}
