+++ VoiceOver Designer/Features/Tests/EditorTests/Doubles/EditorDelegateMock.swift
@@ -1,0 +1,17 @@
+import XCTest
+@testable import Editor
+import Document
+
+class EditorDelegateMock: EditorDelegate {
+    func didDraw(control: A11yDescription) {
+        
+    }
+    
+    func didChangeDocument(controls: [A11yDescription]) {
+        
+    }
+    
+    func didSelect(control: A11yDescription?) {
+        
+    }
+}
