+++ VoiceOver Designer/Features/Tests/EditorTests/NewDrawingTests.swift
@@ -1,0 +1,82 @@
+//
+//  EditorPresenterTests.swift
+//  VoiceOver DesignerTests
+//
+//  Created by Mikhail Rubanov on 05.05.2022.
+//
+
+import XCTest
+@testable import Editor
+import Document
+import DocumentTestHelpers
+
+class NewDrawingTests: EditorAfterDidLoadTests {
+    func test_rectangleDrawnOnTheFly() {
+        sut.mouseDown(on: start10)
+        
+        XCTContext.runActivity(named: "draw on drag") { _ in
+            sut.mouseDragged(on: end60)
+            
+            XCTAssertEqual(drawnControls.first?.frame,
+                           rect10to50, "Draw")
+            
+            XCTAssertNil(documentControls.first, "but not saved yet")
+        }
+        
+        XCTContext.runActivity(named: "saved on release") { _ in
+            sut.mouseUp(on: end60)
+            XCTAssertEqual(drawnControls.first?.frame,
+                           rect10to50)
+        }
+    }
+    
+    func test_drawRectangle_onMouseUp() {
+        sut.mouseDown(on: start10)
+        sut.mouseUp(on: end60)
+        
+        XCTAssertEqual(drawnControls.first?.frame,
+                       rect10to50)
+    }
+    
+    func test_drawSmallerThanMinimalWidth_shouldIncreaseSizeToMinimal_andKeepCenter() {
+        sut.mouseDown(on: start10)
+        sut.mouseUp(on: start10.offset(x: 10, y: 50))
+        
+        XCTAssertEqual(drawnControls.first?.frame,
+                       CGRect(origin: CGPoint(x: 10 + 5 - 44/2,
+                                              y: 10),
+                              size: CGSize(width: 44, height: 50)))
+    }
+    
+    func test_drawSmallerThanMinimalHeight_shouldIncreaseSizeToMinimal_andKeepCenter() {
+        sut.mouseDown(on: start10)
+        sut.mouseUp(on: start10.offset(x: 50, y: 10))
+        
+        XCTAssertEqual(drawnControls.first?.frame,
+                       CGRect(origin: CGPoint(x: 10,
+                                              y: 10 +  5 - 44/2),
+                              size: CGSize(width: 50, height: 44)))
+    }
+    
+    func test_notDrawIfSizeIsSmallerThan5px() {
+        sut.mouseDown(on: start10)
+        sut.mouseUp(on: start10.offset(x: 4, y: 4))
+        
+        XCTAssertNil(drawnControls.first)
+    }
+    
+    func test_drawInReverseDirection() {
+        sut.mouseDown(on: end60)
+        sut.mouseUp(on: start10)
+        
+        XCTAssertEqual(drawnControls.first?.frame,
+                       rect10to50)
+    }
+}
+
+extension CGPoint {
+    func offset(x: CGFloat, y: CGFloat) -> Self {
+        return CGPoint(x: self.x + x,
+                       y: self.y + y)
+    }
+}
